requirements.txt
ipwhois
ipaddress
pythonwhois
tldextract
dnspython
future
